debugpb: init metrics request
