+++ Cargo.toml
@@ -79,6 +79,7 @@ git = "https://github.com/pingcap/rust-rocksdb.git"
 
 [dependencies.kvproto]
 git = "https://github.com/pingcap/kvproto.git"
+branch = "shirly/debug_metrics"
 
 [dependencies.tipb]
 git = "https://github.com/pingcap/tipb.git"
