+++ src/raftstore/store/peer.rs
@@ -1107,6 +1107,7 @@ impl Peer {
                         r.get_cmd_type()
                     ));
                 }
+                _ => unimplemented!(),
             }
 
             if is_read && is_write {
@@ -1647,7 +1648,8 @@ impl Peer {
                 | CmdType::Put
                 | CmdType::Delete
                 | CmdType::DeleteRange
-                | CmdType::Invalid => unreachable!(),
+                | CmdType::Invalid
+                | CmdType::IngestSST => unreachable!(),
             };
             resp.set_cmd_type(cmd_type);
             responses.push(resp);
