+++ src/raftstore/store/worker/apply.rs
@@ -1029,6 +1029,7 @@ impl ApplyDelegate {
                 CmdType::Prewrite | CmdType::Invalid => {
                     Err(box_err!("invalid cmd type, message maybe currupted"))
                 }
+                _ => unimplemented!(),
             }?;
 
             resp.set_cmd_type(cmd_type);
