+++ tests/raftstore/util.rs
@@ -321,6 +321,7 @@ pub fn make_cb(cmd: &RaftCmdRequest) -> (Callback, mpsc::Receiver<RaftCmdRespons
             CmdType::Get | CmdType::Snap => is_read = true,
             CmdType::Put | CmdType::Delete | CmdType::DeleteRange => is_write = true,
             CmdType::Invalid | CmdType::Prewrite => panic!("Invalid RaftCmdRequest: {:?}", cmd),
+            _ => unimplemented!(),
         }
     }
     assert!(is_read ^ is_write, "Invalid RaftCmdRequest: {:?}", cmd);
